Move find module containing node to semantic
Will need it for definition searching when creating debugger mappings

commit-id:6c7bd96d
