+++ crates/cairo-lang-doc/src/parser.rs
@@ -1,7 +1,6 @@
 use std::fmt;
 
 use cairo_lang_debug::DebugWithDb;
-use cairo_lang_defs::db::DefsGroup;
 use cairo_lang_defs::ids::{GenericTypeId, LookupItemId, ModuleId, ModuleItemId, TraitItemId};
 use cairo_lang_diagnostics::DiagnosticsBuilder;
 use cairo_lang_filesystem::db::FilesGroup;
@@ -10,11 +9,9 @@ use cairo_lang_parser::parser::Parser;
 use cairo_lang_semantic::diagnostic::{NotFoundItemType, SemanticDiagnostics};
 use cairo_lang_semantic::expr::inference::InferenceId;
 use cairo_lang_semantic::items::functions::GenericFunctionId;
-use cairo_lang_semantic::items::module::ModuleSemantic;
+use cairo_lang_semantic::lsp_helpers::LspHelpers;
 use cairo_lang_semantic::resolve::{AsSegments, ResolutionContext, ResolvedGenericItem, Resolver};
-use cairo_lang_syntax::node::ast::{Expr, ExprPath, ItemModule};
-use cairo_lang_syntax::node::helpers::GetIdentifier;
-use cairo_lang_syntax::node::{SyntaxNode, TypedSyntaxNode};
+use cairo_lang_syntax::node::ast::{Expr, ExprPath};
 use cairo_lang_utils::Intern;
 use itertools::Itertools;
 use pulldown_cmark::{
@@ -336,7 +333,7 @@ impl<'db> DocumentationCommentParser<'db> {
         path: String,
     ) -> Option<DocumentableItemId<'db>> {
         let syntax_node = item_id.stable_location(self.db)?.syntax_node(self.db);
-        let containing_module = self.find_module_containing_node(&syntax_node)?;
+        let containing_module = self.db.find_module_containing_node(syntax_node)?;
         let mut resolver = Resolver::new(self.db, containing_module, InferenceId::NoContext);
         let mut diagnostics = SemanticDiagnostics::default();
         let segments = self.parse_comment_link_path(path)?;
@@ -372,55 +369,6 @@ impl<'db> DocumentationCommentParser<'db> {
         );
 
         if let Expr::Path(expr_path) = expr { Some(expr_path) } else { None }
-    }
-
-    /// Finds a [`ModuleId`] containing the node.
-    ///
-    /// If the node is located in a virtual file generated by a compiler plugin, this method will
-    /// return the (sub)module of the main, user-written file that leads to the node.
-    fn find_module_containing_node(&self, node: &SyntaxNode<'db>) -> Option<ModuleId<'db>> {
-        let db = self.db;
-
-        // Get the main module of the main file that leads to the node.
-        // The node may be located in a virtual file of a submodule.
-        // This code attempts to get the absolute "parent" of both "module" and "file" parts.
-        let main_module = {
-            // Get the file where the node is located.
-            // This might be a virtual file generated by a compiler plugin.
-            let node_file_id = node.stable_ptr(db).file_id(db);
-
-            // Get the root module of a file containing the node.
-            let node_main_module = db.file_modules(node_file_id).ok()?.iter().copied().next()?;
-
-            // Get the main module of the file.
-            let main_file = db.module_main_file(node_main_module).ok()?;
-
-            // Get the main module of that file.
-            db.file_modules(main_file).ok()?.iter().copied().next()?
-        };
-
-        // Get the stack (bottom-up) of submodule names in the file containing the node, in the main
-        // module, that lead to the node.
-        node.ancestors(db)
-            .filter_map(|node| ItemModule::cast(db, node))
-            .map(|item_module| {
-                item_module
-                    .stable_ptr(db)
-                    .name_green(db)
-                    .identifier(db)
-            })
-            // Buffer the stack to get DoubleEndedIterator.
-            .collect::<Vec<_>>()
-            .into_iter()
-            // And get id of the (sub)module containing the node by traversing this stack top-down.
-            .try_rfold(main_module, |module, name| {
-                let ModuleItemId::Submodule(submodule) =
-                    db.module_item_by_name(module, name).ok()??
-                else {
-                    return None;
-                };
-                Some(ModuleId::Submodule(submodule))
-            })
     }
 }
