+++ crates/cairo-lang-semantic/src/lsp_helpers.rs
@@ -2,11 +2,14 @@ use std::sync::Arc;
 
 use cairo_lang_defs::db::DefsGroup;
 use cairo_lang_defs::ids::{
-    GenericTypeId, ImportableId, LanguageElementId, ModuleId, NamedLanguageElementId,
+    GenericTypeId, ImportableId, LanguageElementId, ModuleId, ModuleItemId, NamedLanguageElementId,
     TraitFunctionId, TraitId,
 };
 use cairo_lang_filesystem::db::{CORELIB_CRATE_NAME, FilesGroup, default_crate_settings};
 use cairo_lang_filesystem::ids::{CrateId, CrateLongId, FileId, SmolStrId, Tracked};
+use cairo_lang_syntax::node::ast::ItemModule;
+use cairo_lang_syntax::node::helpers::GetIdentifier;
+use cairo_lang_syntax::node::{SyntaxNode, TypedSyntaxNode};
 use cairo_lang_utils::Intern;
 use cairo_lang_utils::ordered_hash_map::{Entry, OrderedHashMap};
 use cairo_lang_utils::unordered_hash_set::UnorderedHashSet;
@@ -550,6 +553,47 @@ fn inline_macro_expansion_files_tracked<'db>(
     files
 }
 
+#[salsa::tracked]
+fn find_module_containing_node<'db>(
+    db: &'db dyn Database,
+    _tracked: Tracked,
+    node: SyntaxNode<'db>,
+) -> Option<ModuleId<'db>> {
+    // Get the main module of the main file containing the node.
+    // The node may be located in a virtual file of a submodule.
+    let main_module = {
+        // Get the file where the node is located.
+        // This might be a virtual file generated by a compiler plugin.
+        let node_file_id = node.stable_ptr(db).file_id(db);
+
+        // Get the root module of a file containing the node.
+        *db.file_modules(node_file_id).ok()?.first()?
+    };
+
+    // Get the stack (bottom-up) of submodule names in the file containing the node, in the main
+    // module, that lead to the node.
+    node.ancestors(db)
+        .filter_map(|node| ItemModule::cast(db, node))
+        .map(|item_module| {
+            item_module
+                .stable_ptr(db)
+                .name_green(db)
+                .identifier(db)
+        })
+        // Buffer the stack to get DoubleEndedIterator.
+        .collect::<Vec<_>>()
+        .into_iter()
+        // And get id of the (sub)module containing the node by traversing this stack top-down.
+        .try_rfold(main_module, |module, name| {
+            let ModuleItemId::Submodule(submodule) =
+                db.module_item_by_name(module, name).ok()??
+            else {
+                return None;
+            };
+            Some(ModuleId::Submodule(submodule))
+        })
+}
+
 /// Trait for LSP helpers.
 pub trait LspHelpers<'db>: Database {
     /// Returns all methods in a module that match the given type filter.
@@ -611,6 +655,11 @@ pub trait LspHelpers<'db>: Database {
     /// Returns all files generated by inline macro expansion inside provided module.
     fn inline_macro_expansion_files(&'db self, module_id: ModuleId<'db>) -> &'db Vec<FileId<'db>> {
         inline_macro_expansion_files_tracked(self.as_dyn_database(), (), module_id)
+    }
+
+    /// Finds a [`ModuleId`] containing the node.
+    fn find_module_containing_node(&'db self, node: SyntaxNode<'db>) -> Option<ModuleId<'db>> {
+        find_module_containing_node(self.as_dyn_database(), (), node)
     }
 }
 impl<'db, T: Database + ?Sized> LspHelpers<'db> for T {}
